zed_extension_api

Function download_file

pub fn download_file(
    url: &str,
    file_path: &str,
    file_type: DownloadedFileType,
) -> Result<(), String>

Downloads a file from the given URL and saves it to the given path within the extension’s working directory.

The file will be extracted according to the given file type.