zed_extension_api

Module wit

Modules

_rt 🔒

zed

Structs

CodeLabel

A label containing some code.

CodeLabelSpanLiteral

A span containing a code literal.

KeyValueStore

A key-value store.

Project

A Zed project.

SettingsLocation

Worktree

A Zed worktree.

_RetArea 🔒

Enums

CodeLabelSpan

A span within a code label.

DownloadedFileType

The type of a downloaded file.

LanguageServerInstallationStatus

The installation status for a language server.

Statics

_RET_AREA 🔒

Traits

Guest

Functions

download_file

Downloads a file from the given URL and saves it to the given path within the extension’s working directory.

get_settings

make_file_executable

Makes the file at the given path executable.

set_language_server_installation_status

Updates the installation status for the given language server.

Type Aliases

BuildTaskTemplate

Command

Completion

ContextServerConfiguration

DebugAdapterBinary

DebugConfig

DebugRequest

DebugScenario

DebugTaskDefinition

EnvVars

Range

ResolvedTask

SlashCommand

SlashCommandArgumentCompletion

SlashCommandOutput

StartDebuggingRequestArgumentsRequest

Symbol