zed_extension_api::wit

Module _rt

Modules

alloc

Memory allocation APIs

Structs

Resource

A type which represents a component model resource, either imported or exported into this component.

String

A UTF-8–encoded, growable string.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Traits

AsI32

AsI64

WasmResource

A trait which all wasm resources implement, namely providing the ability to drop a resource.

Functions

as_i32

as_i64

bool_lift⚠

cabi_dealloc⚠

invalid_enum_discriminant⚠

run_ctors_once

string_lift⚠