zed_extension_api::wit

Type Alias SlashCommandArgumentCompletion

pub type SlashCommandArgumentCompletion = SlashCommandArgumentCompletion;

Aliased Type

pub struct SlashCommandArgumentCompletion {
    pub label: String,
    pub new_text: String,
    pub run_command: bool,
}

Fields
label: String

The label to display for this completion.

new_text: String

The new text that should be inserted into the command when this completion is accepted.

run_command: bool

Whether the command should be run when accepting this completion.