zed_extension_api::wit::zed::extension

Module dap

Structs

AttachRequest

BuildTaskDefinitionTemplatePayload

DebugAdapterBinary

The lowest-level representation of a debug session, which specifies:

DebugConfig

Debug Config is the “highest-level” configuration for a debug session. It comes from a new process modal UI; thus, it is essentially debug-adapter-agnostic. It is expected of the extension to translate this generic configuration into something that can be debugged by the adapter (debug scenario).

DebugScenario

Debug Scenario is the user-facing configuration type (used in debug.json). It is still concerned with what to debug and not necessarily how to do it (except for any debug-adapter-specific configuration options).

DebugTaskDefinition

LaunchRequest

StartDebuggingRequestArguments

TaskTemplate

TcpArguments

TcpArgumentsTemplate

Enums

BuildTaskDefinition

DebugRequest

StartDebuggingRequestArgumentsRequest

Functions

resolve_tcp_template

Resolves a specified TcpArgumentsTemplate into TcpArguments

Type Aliases

BuildTaskTemplate

A task template for building a debug target.

EnvVars

ResolvedTask

A task template with substituted task variables.